chore: migrate auth package vitest
