--- packages/auth/jest.config.js
@@ -1,10 +1,0 @@
-const config = require('../../jest/config');
-
-module.exports = Object.assign({}, config, {
-  coverageThreshold: {
-    global: {
-      // FIXME: increase to 90
-      lines: 30,
-    },
-  },
-});
